refactor: update model imports and fix lint errors
