+++ src/app/order/history/history.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { OrderService, Order } from '../order.service';
+import { Order } from '../order.model';
+import { OrderService } from '../order.service';
 import { Config } from '../../restaurant/restaurant.service';
 import * as io from 'socket.io-client';
 import { environment } from '../../../environments/environment';
