+++ src/app/shared/item-total.pipe.ts
@@ -1,5 +1,5 @@
 import {Pipe, PipeTransform} from '@angular/core';
-import {Item} from "../order/order.service";
+import {Item} from '../order/order.model';
 
 
 @Pipe({
